// frontend/src/api/schema.d.ts
export interface components {
    schemas: {
// ...
        TokenResponse: {
            access_token: string;
            refresh_token: string;
            /** @default bearer */
            token_type: string;
            expires_in: number;
        };
        /** UserResponse */
        UserResponse: {
            /** Format: uuid */
            id: string;
            /** Format: email */
            email: string;
            full_name: string;
            is_active: boolean;
            /** Format: date-time */
            created_at: string;
        };
// ...
    };
// ...
}